Also reorg some synthesis tests
